fix: fixing some `pipeline.py` and `errors.py` inconsistencies and respective tests  (#277)
* initial import

* cleaning init

* fixing errors.py duplication

* fixing linting

* fixing pipeline run time error

* fixing pipeline run time error

* fixing pipeline run time tests
