+++ haystack_experimental/core/__init__.py
@@ -5,4 +5,4 @@
 from .pipeline.base import ComponentPriority, PipelineBase
 from .super_component import SuperComponent
 
-_all_ = ["ComponentPriority", "PipelineBase", "SuperComponent"]
+_all_ = ["PipelineBase"]
